Update portfolio holdings to support decimal quantities and enhance CSV upload functionality. Modify database schema and models to use NUMERIC type for quantity, update API endpoints for handling decimal values, and improve CSV parsing to accommodate different delimiters and decimal formats.

+++ backend/migrate_add_portfolio_holdings.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS portfolio_holdings (
     ticker VARCHAR(20),
     name VARCHAR(255) NOT NULL,
     purchase_date TIMESTAMP NOT NULL,
-    quantity INTEGER NOT NULL,
+    quantity NUMERIC(15, 6) NOT NULL,
     purchase_price VARCHAR(50) NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
